converted points to use doubles

+++ src/model/Point.scala
@@ -3,9 +3,9 @@ package model
 /**
   * Created by agieg on 4/19/2017.
   */
-class Point(_x: Int, _y: Int) {
-  def x: Int = _x
-  def y: Int = _y
+class Point(_x: Double, _y: Double ){
+  def x: Double = _x
+  def y: Double = _y
 
   override def toString: String = {
     "("+this.x+", "+this.y+")"
